[Gold V] Title: Fly me to the Alpha Centauri, Time: 0 ms, Memory: 3684 KB -BaekjoonHub

+++ 백준/Gold/1011. Fly me to the Alpha Centauri/Fly me to the Alpha Centauri.cc
@@ -5,7 +5,7 @@
 
 using namespace std;
 
-long long t, x, y, diff;
+int t, x, y, diff;
 vector<long long> v(1, 0);
 
 int main()
